EPICS_CAS_IGNORE_NET_LIST lets servers ignore requests from whole subnets.
An EPICS channel access server will ignore requests from IP addresses that are
from one of the IP subnets specified in EPICS_CAS_IGNORE_NET_LIST.

The subnet specifications in EPICS_CAS_IGNORE_NET_LIST must be separated with
spaces.

Three forms of subnet specifications are supported:

- CIDR notation, example: 192.168.12.0/24
- Address and mask notation, example: 192.168.12.0:255.255.255.0
- Single IP address, example: 192.168.12.10

Notes for the dependency from EPICS base:

This version of PCAS only supports EPICS_CAS_IGNORE_NET_LIST if it is compiled
against an EPICS base that supports this feature too. It is however, backwards
compatible and can be compiled against older versions of EPICS base. In this
case, EPICS_CAS_IGNORE_NET_LIST is not supported.

Notes for software that uses PCAS:

This patch adds a private element to casDGIntfIO defined in casDGIntfIO.h.
It shouldn't be necessary to change the source code of applications
that are compiled against PCAS, but they need to be recompiled.