Access security now supports IP address access groups (IPAG).
Up to now, channel access security allowed to specify access rules for clients
based on user- and host names.

This patch adds IP access groups. Now access rules may contain the client's IP
address too. The IP access group, "IPAG", can be used together with host access
groups, "HAG" or can replace these.

Notes for the dependency from EPICS base:

This version of PCAS only supports IP access groups if it is compiled
against an EPICS base that supports this feature too. It is however, backwards
compatible and can be compiled against older versions of EPICS base. In this
case, IP access groups are not supported.

Notes for software that uses PCAS:

This patch changes the API of PCAS. In general, users of the library have to
provide the IP address of the client. These are the changes:

- In casChannelI.h, casChannelI::setOwner has an additional parameter
- In casPVI.h, casPVI::createChannel has an additional parameter
- In casStrmClient.h, class casStrmClient has a new private element
- In casdef.h,  casPV::createChannel has an additional parameter
- In casdef.h,  casChannelI::setOwner has an additional parameter

The changes are active when EPICS_HAS_AS_IPAG is defined, which is the case
when the EPICS base also supports IP access groups.